using kind of block bfs, some patches of bfs to initialize the genes

+++ src/renderer.h
@@ -20,7 +20,7 @@ class renderer {
 
 	double SCALE;
 	bool scale_edited;
-	int FPS;
+	double FPS;
 	float draw_decimation;
 	ALLEGRO_FONT *debug_font;
 	bool step_it;
